+++ packages/fuels-test-helpers/src/fuel_bin_service.rs
@@ -1,8 +1,10 @@
 use std::{
     net::{IpAddr, SocketAddr},
+    path::PathBuf,
     time::Duration,
 };
 
+use fuel_core_chain_config::{ChainConfig, SnapshotWriter, StateConfig};
 use fuel_core_client::client::FuelClient;
 use fuel_core_services::State;
 use fuels_core::{
@@ -10,9 +12,96 @@ use fuels_core::{
     types::errors::{Error, Result as FuelResult},
 };
 use portpicker::{is_free, pick_unused_port};
+use tempfile::{tempdir, TempDir};
 use tokio::{process::Command, spawn, task::JoinHandle, time::sleep};
 
-use crate::{node_types::NodeConfig, ExtendedConfig};
+use crate::node_types::{DbType, NodeConfig, Trigger};
+
+#[derive(Debug)]
+pub(crate) struct ExtendedConfig {
+    pub node_config: NodeConfig,
+    pub chain_config: ChainConfig,
+    pub state_config: StateConfig,
+    pub snapshot_dir: TempDir,
+}
+
+impl ExtendedConfig {
+    pub fn args_vec(&self) -> fuels_core::types::errors::Result<Vec<String>> {
+        let port = self.node_config.addr.port().to_string();
+        let mut args = vec![
+            "run".to_string(),
+            "--ip".to_string(),
+            "127.0.0.1".to_string(),
+            "--port".to_string(),
+            port,
+            "--snapshot".to_string(),
+            self.snapshot_dir
+                .path()
+                .to_str()
+                .expect("Failed to find config file")
+                .to_string(),
+        ];
+
+        args.push("--db-type".to_string());
+        match &self.node_config.database_type {
+            DbType::InMemory => args.push("in-memory".to_string()),
+            DbType::RocksDb(path_to_db) => {
+                args.push("rocks-db".to_string());
+                let path = path_to_db.as_ref().cloned().unwrap_or_else(|| {
+                    PathBuf::from(std::env::var("HOME").expect("HOME env var missing"))
+                        .join(".fuel/db")
+                });
+                args.push("--db-path".to_string());
+                args.push(path.to_string_lossy().to_string());
+            }
+        }
+
+        if let Some(cache_size) = self.node_config.max_database_cache_size {
+            args.push("--max-database-cache-size".to_string());
+            args.push(cache_size.to_string());
+        }
+
+        match self.node_config.block_production {
+            Trigger::Instant => {
+                args.push("--poa-instant=true".to_string());
+            }
+            Trigger::Never => {
+                args.push("--poa-instant=false".to_string());
+            }
+            Trigger::Interval { block_time } => {
+                args.push(format!(
+                    "--poa-interval-period={}ms",
+                    block_time.as_millis()
+                ));
+            }
+        };
+
+        args.extend(
+            [
+                (self.node_config.vm_backtrace, "--vm-backtrace"),
+                (self.node_config.utxo_validation, "--utxo-validation"),
+                (self.node_config.debug, "--debug"),
+            ]
+            .into_iter()
+            .filter(|(flag, _)| *flag)
+            .map(|(_, arg)| arg.to_string()),
+        );
+
+        Ok(args)
+    }
+
+    pub fn write_temp_snapshot_files(self) -> FuelResult<TempDir> {
+        let mut writer = SnapshotWriter::json(self.snapshot_dir.path());
+        writer
+            .write_chain_config(&self.chain_config)
+            .map_err(|e| error!(Other, "could not write chain config: {}", e))?;
+        writer
+            .write_state_config(self.state_config)
+            .map_err(|e| error!(Other, "could not write state config: {}", e))?;
+
+        Ok(self.snapshot_dir)
+    }
+}
 
 pub struct FuelService {
     pub bound_address: SocketAddr,
@@ -20,23 +109,29 @@ pub struct FuelService {
 }
 
 impl FuelService {
-    pub async fn new_node(config: ExtendedConfig) -> FuelResult<Self> {
-        let requested_port = config.node_config.addr.port();
+    pub async fn new_node(
+        node_config: NodeConfig,
+        chain_config: ChainConfig,
+        state_config: StateConfig,
+    ) -> FuelResult<Self> {
+        let requested_port = node_config.addr.port();
 
         let bound_address = match requested_port {
             0 => get_socket_address()?,
-            _ if is_free(requested_port) => config.node_config.addr,
+            _ if is_free(requested_port) => node_config.addr,
             _ => return Err(Error::IO(std::io::ErrorKind::AddrInUse.into())),
         };
 
         let node_config = NodeConfig {
             addr: bound_address,
-            ..config.node_config
+            ..node_config
         };
 
         let extended_config = ExtendedConfig {
             node_config,
-            ..config
+            state_config,
+            chain_config,
+            snapshot_dir: tempdir()?,
         };
 
         let addr = extended_config.node_config.addr;
